refactor(llm): extract openai_compatible behind the Codec trait
Move the OpenAI Chat Completions wire logic out of the adapter into
codec/openai_compatible/ as an `impl Codec`: wire types (wire.rs), pure
mapping (translate.rs), request encoding (request.rs), response decoding
(response.rs), and the streaming decoder (stream.rs). The codec is pure
— no HTTP, auth, or base URL.

providers/openai_compatible.rs becomes a thin transport shell: it keeps
the struct, builders, auth, and validate_request, owns the streaming
byte loop + SSE `data:` framing, and delegates all translation to the
codec. The two hand-rolled stream unfolds collapse into one. Public
constructors (`new`/`with_name`/`with_catalog`/…) are unchanged —
fabro-agent's parity_matrix and the wire snapshot suite pass unmodified.

Extracting the first real codec refined the trait signatures (kept in
codec/mod.rs): the canonical `Request` moves into `CodecCtx` (decoders
need it for tool-argument parsing and the stream model fallback), and
the header-parsed `rate_limit` threads into `decode_response` /
`stream_decoder`. `on_event` returns `Result` for dialect error events.

The 29 in-module unit tests move into the codec submodules alongside
the code they exercise. Behavior is unchanged: all 515 fabro-llm tests
(incl. the 126 wire snapshots) pass without snapshot edits.
